requirements-base.txt
# 기본 HTTP 및 웹 관련
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
lxml[html_clean]>=4.9.0

# RSS 및 뉴스 수집
feedparser>=6.0.0
newspaper3k>=0.2.8

# 데이터베이스 및 캐시
redis>=5.0.0
pymongo>=4.5.0
pinecone>=2.2.0
supabase>=2.0.0

# 알림 시스템 관련
schedule>=1.2.0
sib-api-v3-sdk>=6.0.0

# 날짜 및 시간 처리
python-dateutil>=2.8.0

# 설정 및 데이터 처리
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# 웹 프레임워크
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 비동기 처리
celery>=5.3.0
eventlet>=0.33.0

# 인증 및 보안
certifi>=2023.0.0

# 추가 유틸리티
pathlib2>=2.3.0

# OpenAI (notification-system에서 사용)
openai>=1.0.0

# Protobuf (transformers 모델 로딩에 필요)
protobuf>=4.21.0 